[TASK] Prepare v9/v10 compatible release
* Switch RDCT parameter evaluation to a PSR-15 middleware
* Raise dependency to TYPO3 v9+
