+++ Classes/Redirects.php
@@ -19,89 +19,10 @@ use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
- * Evaluates &RDCT parameter in the frontend
- * this is done through ->sendRedirect()
- * and was previously located in $TSFE->sendRedirect()
- *
- * For creating entries, use the method ->makeRedirectUrl()
+ * Creates Redirect entries
  */
 class Redirects
 {
-    /**
-     * Looks up the value of $this->RDCT in the database and if it is
-     * found to be associated with a redirect URL then the redirection
-     * is carried out with a 'Location:' header
-     * May exit after sending a location-header.
-     */
-    public function sendRedirect()
-    {
-        // TSFE instantiated from BE or CLI context, nothing to do
-        if (!(TYPO3_REQUESTTYPE & TYPO3_REQUESTTYPE_FE)) {
-            return;
-        }
-
-        $redirectHash = GeneralUtility::_GP('RDCT');
-        // No GET parameter set, do nothing
-        if (empty($redirectHash)) {
-            return;
-        }
-
-        // Website is unavailable, do nothing
-        if ($GLOBALS['TYPO3_CONF_VARS']['FE']['pageUnavailable_force']
-            && !GeneralUtility::cmpIP(
-                GeneralUtility::getIndpEnv('REMOTE_ADDR'),
-                $GLOBALS['TYPO3_CONF_VARS']['SYS']['devIPmask']
-            )
-        ) {
-            return;
-        }
-
-        $row = $this->fetchRedirectRecord($redirectHash);
-        if (is_array($row)) {
-            $this->updateMD5paramsRecord($redirectHash);
-            header('Location: ' . $row['params']);
-            die;
-        }
-    }
-
-    /**
-     * Checks cache_md5params if a redirect hash is available
-     *
-     * @param string $redirectHash
-     * @return mixed
-     */
-    protected function fetchRedirectRecord(string $redirectHash)
-    {
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-            ->getQueryBuilderForTable('cache_md5params');
-        return $queryBuilder
-            ->select('params')
-            ->from('cache_md5params')
-            ->where(
-                $queryBuilder->expr()->eq(
-                    'md5hash',
-                    $queryBuilder->createNamedParameter($redirectHash, \PDO::PARAM_STR)
-                )
-            )
-            ->execute()
-            ->fetch();
-    }
-
-    /**
-     * Updates the tstamp field of a cache_md5params record to the current time.
-     *
-     * @param string $redirectHash The hash string identifying the cache_md5params record for which to update the "tstamp" field to the current time.
-     */
-    protected function updateMD5paramsRecord(string $redirectHash)
-    {
-        $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable('cache_md5params');
-        $connection->update(
-            'cache_md5params',
-            ['tstamp' => $GLOBALS['EXEC_TIME']],
-            ['md5hash' => $redirectHash]
-        );
-    }
-
     /**
      * Create a shortened "redirect" URL with specified length from an incoming URL
      *
