+++ LICENSE.txt
@@ -1,9 +1,3 @@
-Some icons used in the TYPO3 project are retrieved from the "Silk" icon set of
-Mark James, which can be found at http://famfamfam.com/lab/icons/silk/. This
-set is distributed under a Creative Commons Attribution 2.5 License. The
-license can be found at http://creativecommons.org/licenses/by/2.5/.
----------------------------------
-
                     GNU GENERAL PUBLIC LICENSE
                        Version 2, June 1991
