+++ composer.json
@@ -12,12 +12,12 @@
 		}
 	],
 	"require": {
-		"php": "^7.0",
-		"typo3/cms-core": ">=8.7.0 <=9.9.99",
-		"typo3/cms-frontend": ">=8.7.0 <=9.9.99"
+		"php": "^7.2",
+		"typo3/cms-core": ">=9.5.0 <=10.9.99",
+		"typo3/cms-frontend": ">=9.5.0 <=10.9.99"
 	},
 	"replace": {
-		"rdct": "*"
+		"typo3-ter/rdct": "*"
 	},
 	"extra": {
 		"typo3/cms": {
